Cleaned up the project from unnecessery code
