+++ src/main/java/messagequeue/RMIMessageQueueImpl.java
@@ -10,54 +10,17 @@ import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 
 public class RMIMessageQueueImpl extends java.rmi.server.UnicastRemoteObject implements RMIMessageQueue {
-
-	BlockingQueue<String> queueStrings;
-
 	BlockingQueue<Attacks.BasicWebAttack> queue;
-
-	Queue<Worker> workers;
 	
 	public RMIMessageQueueImpl()  throws RemoteException {
 		super();
-		queueStrings = new LinkedBlockingQueue<>();
         queue = new LinkedBlockingQueue<>();
-		workers = new LinkedList<>();
-	}
-	
-	@Override
-	public void createTask(String task) throws RemoteException {
-		System.out.println("Received task: "+task);
-		queueStrings.add(task);
-
-//		Worker w = workers.remove();
-//        try {
-//            w.handleTask(queueStrings.take());
-//        } catch (InterruptedException e) {
-//            e.printStackTrace();
-//        }
-//        workers.add(w);
 	}
 
 	@Override
 	public void createTask(Attacks.BasicWebAttack attack) throws RemoteException {
-//        System.out.println("Received task: "+task);
+        System.out.println("Received task: ");
         queue.add(attack);
-	}
-
-
-	public void registerWorker(Worker w) throws RemoteException {
-		workers.add(w);
-		System.out.println("Added new worker");
-	}
-
-	@Override
-	public String getTaskString() throws RemoteException {
-        try {
-            return(queueStrings.take());
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-        return null;
 	}
 
 	@Override
