--- src/main/java/util/AttackType.java
@@ -1,5 +1,0 @@
-package util;
-
-public enum AttackType {
-    SQLINJECTION, PHPINJECTION, BRUTEFORCE;
-}
