+++ src/main/java/util/UrlAttacker.java
@@ -1,4 +1,4 @@
-package application;
+package util;
 
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
@@ -8,8 +8,6 @@ import org.apache.http.client.entity.UrlEncodedFormEntity;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.message.BasicNameValuePair;
-import util.Attack;
-import util.AttackType;
 import util.Attacks;
 
 import java.io.IOException;
@@ -91,46 +89,6 @@ public class UrlAttacker {
     }
 
 
-    // VERSION 1 <-----------
-    private static List<List<NameValuePair>> bruteforceAttack(Attack attack){
-        String url = attack.url;
-        List<List<NameValuePair>> successfulParamSets = new ArrayList<>();
-
-        if(url == null || attack.type==null) { return null; }
-
-        HttpClient httpclient = HttpClients.createDefault();
-
-        for (List<NameValuePair> paramsSet : attack.paramsBatch) {
-            HttpPost httppost = constructRequest(attack.url, paramsSet);
-            String response = getResponse(httpclient, httppost);
-
-            if(response != null && response.contains(attack.successIdentifier)){
-                System.out.println("SUCCESS!!!!");
-                //System.out.println(response);
-                successfulParamSets.add(paramsSet);
-            }
-        }
-        return successfulParamSets;
-    }
-
-    public static List<List<NameValuePair>> performAttack(Attack attack){
-        if(attack.type == AttackType.BRUTEFORCE){
-            return bruteforceAttack(attack);
-        }
-        else if(attack.type == AttackType.PHPINJECTION){
-            // call a method
-        }
-        else if(attack.type == AttackType.SQLINJECTION){
-            // call a method
-        }
-
-        return null;
-    }
-    // ---------> VERSION 1
-
-
-    // VERSION 2 <----------
-
     public static List<List<NameValuePair>> performAttack(Attacks.BasicWebAttack attack){
         if(attack instanceof Attacks.BruteforceAttack){
             return bruteforceAttack((Attacks.BruteforceAttack) attack);
@@ -156,50 +114,5 @@ public class UrlAttacker {
             }
         }
         return successfulParamSets;
-    }
-
-    // -----------> VERSION 2
-
-    public static void main(String[] args) {
-        List<NameValuePair> params = new ArrayList<NameValuePair>(2);
-        params.add(new BasicNameValuePair("param-1", "12345"));
-        params.add(new BasicNameValuePair("field", "fuzzed"));
-
-        List<NameValuePair> params2 = new ArrayList<NameValuePair>(2);
-        params2.add(new BasicNameValuePair("username", "username"));
-        params2.add(new BasicNameValuePair("password", "passwor"));//missing 'd'
-
-        List<NameValuePair> params3 = new ArrayList<NameValuePair>(3);
-        params3.add(new BasicNameValuePair("username", "username"));
-        params3.add(new BasicNameValuePair("password", "password"));
-        params3.add(new BasicNameValuePair("age", "22a")); //not only numbers
-
-        List<List<NameValuePair>> paramsBatch = new ArrayList<>();
-        paramsBatch.add(params);
-        paramsBatch.add(params2);
-        paramsBatch.add(params3);
-
-        List<NameValuePair> credentials = new ArrayList<>(2);
-        credentials.add(new BasicNameValuePair("username", "username"));
-        credentials.add(new BasicNameValuePair("password", "password"));
-
-        paramsBatch.add(credentials);
-
-        Attack attack = new Attack(AttackType.BRUTEFORCE, "http://localhost:5000/login", credentials, paramsBatch, "Internal Server Error");
-        List<List<NameValuePair>> attackRes = performAttack(attack);
-        if(attackRes != null)
-            System.out.println(attackRes.toString());
-        else
-            System.out.println("Something went wrong :(");
-
-
-        Attacks.BruteforceAttack bfAttack = new Attacks.BruteforceAttack("http://localhost:5000/login",paramsBatch, "Internal Server Error");
-        attackRes = performAttack(bfAttack);
-        if(attackRes != null)
-            System.out.println(attackRes.toString());
-        else
-            System.out.println("Something went wrong :(");
-
-
     }
 }
